refactor(sync): split sync-service into facade and cohesive modules
