+++ src/main/sync/sync-service.ts
@@ -1,1707 +1,107 @@
 // SPDX-License-Identifier: GPL-2.0-or-later
-// Sync orchestration: bundling, conflict resolution, debounce upload, before-quit flush
+// Sync orchestration: bundling, conflict resolution, debounce upload,
+// before-quit flush.
+//
+// This file is the facade: it owns no logic of its own (besides
+// `_resetForTests`, which delegates to each sibling module's own reset)
+// and re-exports the full public surface from the sibling modules in
+// this directory:
+//
+//   sync-runtime-state.ts    — shared mutable state + small accessors
+//   sync-scope.ts            — SyncScope matching / download filtering
+//   sync-password.ts         — credentials + password-check validation
+//   sync-merge-dispatch.ts   — per-sync-unit upload/merge/dispatch
+//   sync-scan.ts             — remote data inspection (scan/undecryptable)
+//   sync-typing-remote.ts    — typing-analytics remote day bookkeeping
+//   sync-execute.ts          — full-pass download/upload sync
+//   sync-polling.ts          — 3-minute background polling
+//   sync-analytics.ts        — Analyze-panel-triggered analytics sync
+//   sync-flush.ts            — debounced auto-sync + before-quit handler
+//
+// New sync logic belongs in the sibling module whose responsibility it
+// extends — not here. External consumers (sync-ipc.ts, main/index.ts,
+// the 9+ store files that call `notifyChange`, and every test file's
+// whole-module mock of this facade path) must keep importing this
+// facade path, never a submodule directly. See
+// .claude/tasks/backlog/Task-split-sync-service.md and
+// .claude/rules/file-splitting.md for the split rationale.
+//
+// New module-private mutable state must either live on `syncRuntime`
+// (sync-runtime-state.ts) or ship its own `*ForTests` reset seam called
+// from `_resetForTests` below — the same convention
+// typing-analytics-service.ts's facade split uses.
 
-import { app } from 'electron'
-import { join } from 'node:path'
-import { readFile, writeFile, mkdir, readdir } from 'node:fs/promises'
-import { encrypt, decrypt, retrievePasswordResult, storePassword, clearPassword } from './sync-crypto'
-import { loadAppConfig } from '../app-config'
-import { getAuthStatus } from './google-auth'
-import {
-  listFiles,
-  downloadFile,
-  uploadFile,
-  deleteFile,
-  driveFileName,
-  driveFilenamePrefix,
-  syncUnitFromFileName,
-  type DriveFile,
-} from './google-drive'
-import { unlink } from 'node:fs/promises'
-import { pLimit } from '../../shared/concurrency'
-import { IpcChannels } from '../../shared/ipc/channels'
-import { mergeEntries, gcTombstones, safeTimestamp, MalformedSyncBundleError, type EntryMeta } from './merge'
-import {
-  mergePackIndexBundle,
-  mergePackBodyBundle,
-  parsePackBodySyncUnit,
-  packBodyLocalWins,
-  pinPackBodyMtimeAfterUpload,
-  statPackBodyLocalMtime,
-} from './pack-bundle-merge'
-import {
-  readIndexFile,
-  bundleSyncUnit,
-  collectAllSyncUnits,
-  collectAnalyticsSyncUnitsForUid,
-  isAnalyticsSyncUnit,
-  isRunLogSyncUnit,
-} from './sync-bundle'
-import { isSafePathSegment } from '../utils/safe-filename'
-import { MAX_RUN_LOGS_PER_KEYBOARD } from '../../shared/types/typing-run-log'
-import {
-  applyRemoteKeyboardMetaIndex,
-  backfillKeyboardMeta,
-  getActiveKeyboardMetaMap,
-  readKeyboardMetaIndex,
-} from './keyboard-meta'
-import { KEYBOARD_META_SYNC_UNIT, type KeyboardMetaIndex } from '../../shared/types/keyboard-meta'
-import { KEY_LABEL_SYNC_UNIT } from '../key-label-store'
-import { TYPING_TEST_TEXT_SYNC_UNIT } from '../typing-test-text-store'
-import { I18N_SYNC_UNIT_PREFIX, I18N_INDEX_SYNC_UNIT } from '../../shared/types/i18n-store'
-import { THEME_SYNC_UNIT_PREFIX, THEME_INDEX_SYNC_UNIT } from '../../shared/types/theme-store'
-import { broadcastToAllWindows } from '../utils/broadcast'
-import { runPackGcAfterPass } from './pack-gc'
-import {
-  parseTypingAnalyticsDeviceDaySyncUnit,
-  typingAnalyticsDeviceDaySyncUnit,
-} from '../typing-analytics/sync'
-import { applyRowsToCache } from '../typing-analytics/jsonl/apply-to-cache'
-import { readRows } from '../typing-analytics/jsonl/jsonl-reader'
-import {
-  deviceDayDir,
-  deviceDayJsonlPath,
-  listDeviceDays,
-  readPointerKey,
-} from '../typing-analytics/jsonl/paths'
-import { utcDayBoundaryMs, type UtcDay } from '../typing-analytics/jsonl/utc-day'
-import { getTypingAnalyticsDB } from '../typing-analytics/db/typing-analytics-db'
-import { getMachineHash } from '../typing-analytics/machine-hash'
-import {
-  emptySyncState,
-  isReconcilePending,
-  loadSyncState,
-  saveSyncState,
-  type TypingSyncState,
-} from '../typing-analytics/sync-state'
-import { log } from '../logger'
-import type { SyncBundle, SyncProgress, SyncEnvelope, UndecryptableFile, SyncDataScanResult, SyncScope, SyncCredentialFailureReason, SyncCredentialResult, SyncExecuteStatus, SyncSkipReason } from '../../shared/types/sync'
-import { syncCredentialI18nKey } from '../../shared/types/sync'
+import { syncRuntime } from './sync-runtime-state'
+import { stopPolling } from './sync-polling'
+import { clearQuitFinalizersForTests } from './sync-flush'
 
-export class SyncCredentialError extends Error {
-  readonly reason: SyncCredentialFailureReason
-  constructor(reason: SyncCredentialFailureReason, namespace: 'readiness' | 'changePasswordError' = 'changePasswordError') {
-    super(syncCredentialI18nKey(namespace, reason))
-    this.reason = reason
+// --- Test helpers -------------------------------------------------------
+
+export function _resetForTests(): void {
+  if (syncRuntime.debounceTimer) {
+    clearTimeout(syncRuntime.debounceTimer)
+    syncRuntime.debounceTimer = null
   }
+  stopPolling()
+  syncRuntime.pendingChanges.clear()
+  syncRuntime.lastKnownRemoteState.clear()
+  syncRuntime.isSyncing = false
+  syncRuntime.isQuitting = false
+  syncRuntime.progressCallback = null
+  syncRuntime.passwordCheckValidated = false
+  clearQuitFinalizersForTests()
 }
 
-const SYNC_CONCURRENCY = 10
-const DEBOUNCE_MS = 10_000
-const POLL_INTERVAL_MS = 3 * 60 * 1000 // 3 minutes
-const PASSWORD_CHECK_UNIT = 'password-check'
-const PASSWORD_CHECK_PAYLOAD = JSON.stringify({ type: 'password-check', version: 1 })
+// --- Public re-exports ---------------------------------------------------
+// Explicit named re-exports only (never `export *`) so the facade's public
+// surface is grep-able in one place and stays byte-identical to what it was
+// before the split.
 
-type ProgressCallback = (progress: SyncProgress) => void
+export { SyncCredentialError } from './sync-password'
 
-let debounceTimer: ReturnType<typeof setTimeout> | null = null
-let pendingChanges = new Set<string>()
-let progressCallback: ProgressCallback | null = null
-let isQuitting = false
-let isSyncing = false
-let pollTimer: ReturnType<typeof setInterval> | null = null
-let passwordCheckValidated = false
-const lastKnownRemoteState = new Map<string, string>() // fileName -> modifiedTime
+export {
+  hasPendingChanges,
+  cancelPendingChanges,
+  isSyncInProgress,
+  setProgressCallback,
+} from './sync-runtime-state'
 
-export function hasPendingChanges(): boolean {
-  return pendingChanges.size > 0
-}
-
-export function cancelPendingChanges(prefix?: string): void {
-  if (prefix) {
-    for (const unit of pendingChanges) {
-      if (unit.startsWith(prefix)) pendingChanges.delete(unit)
-    }
-  } else {
-    pendingChanges.clear()
-  }
-  if (pendingChanges.size === 0 && debounceTimer) {
-    clearTimeout(debounceTimer)
-    debounceTimer = null
-  }
-  broadcastPendingStatus()
-}
-
-export function isSyncInProgress(): boolean {
-  return isSyncing
-}
-
-function broadcastPendingStatus(): void {
-  broadcastToAllWindows(IpcChannels.SYNC_PENDING_STATUS, hasPendingChanges())
-}
-
-export function setProgressCallback(cb: ProgressCallback): void {
-  progressCallback = cb
-}
-
-function emitProgress(progress: SyncProgress): void {
-  progressCallback?.(progress)
-}
-
-function errorMessage(err: unknown, fallback: string): string {
-  return err instanceof Error ? err.message : fallback
-}
-
-export function matchesScope(syncUnit: string | null, scope: SyncScope): boolean {
-  if (scope === 'all') return true
-  if (syncUnit === null) return false
-  // 'packs' is checked before the unconditional keyboard-meta/key-label/
-  // typing-test-text `true`s below — otherwise a 'packs'-scoped download
-  // would also pull those unrelated global units in, since their own
-  // checks don't otherwise care what scope was asked for. Returning here
-  // also means 'packs' never falls through to the i18n/themes rejection
-  // further down, and never matches anything else (favorites/keyboards).
-  // See .claude/docs/DATA-INVENTORY.md §3.7's packs-scope definition for
-  // the full discovery-route writeup this ordering exists to support.
-  if (scope === 'packs') {
-    return syncUnit.startsWith(I18N_SYNC_UNIT_PREFIX) || syncUnit.startsWith(THEME_SYNC_UNIT_PREFIX)
-  }
-  if (syncUnit === KEYBOARD_META_SYNC_UNIT) return true // meta follows every scope
-  if (syncUnit === KEY_LABEL_SYNC_UNIT) return true // key-labels follow every scope (global, all-keyboard)
-  if (syncUnit === TYPING_TEST_TEXT_SYNC_UNIT) return true // imported typing-test texts follow every scope (global, all-keyboard)
-  // i18n/themes only match 'all' or 'packs' (both short-circuited above) —
-  // excluded from favorites/keyboard-scoped syncs. NOT reliably discovered
-  // by the 3-minute poll alone despite it using 'all': polling only merges
-  // a CHANGED modifiedTime since its own last snapshot, so a pack already
-  // on Drive before this machine's first poll is invisible to it forever.
-  // See matchesScope's doc in .claude/docs/DATA-INVENTORY.md §3.7 (and
-  // .claude/rules/settings-persistence.md) for the full writeup and the
-  // 'packs'-scope discovery routes that actually close this gap.
-  if (syncUnit.startsWith(I18N_SYNC_UNIT_PREFIX) || syncUnit.startsWith(THEME_SYNC_UNIT_PREFIX)) return false
-  if (scope === 'favorites') return syncUnit.startsWith('favorites/')
-  if (typeof scope === 'object' && 'favorites' in scope) {
-    return syncUnit.startsWith('favorites/') || syncUnit.startsWith(`keyboards/${scope.keyboard}/`)
-  }
-  return syncUnit.startsWith(`keyboards/${scope.keyboard}/`)
-}
+export { matchesScope, shouldDownloadSyncUnit } from './sync-scope'
 
 // Re-export the analytics/run-log sync-unit detectors so existing
 // callers (sync-ipc, tests) keep importing them from sync-service.
-export { isAnalyticsSyncUnit, isRunLogSyncUnit }
-
-async function listLocalKeyboardUids(): Promise<Set<string>> {
-  const userData = app.getPath('userData')
-  const keyboardsDir = join(userData, 'sync', 'keyboards')
-  const uids = new Set<string>()
-  try {
-    const entries = await readdir(keyboardsDir, { withFileTypes: true })
-    for (const entry of entries) {
-      if (entry.isDirectory()) uids.add(entry.name)
-    }
-  } catch { /* dir doesn't exist */ }
-  return uids
-}
-
-export function shouldDownloadSyncUnit(
-  syncUnit: string | null,
-  scope: SyncScope,
-  localKeyboardUids: Set<string>,
-): boolean {
-  if (!syncUnit) return false
-  if (!matchesScope(syncUnit, scope)) return false
-  // Keyboard-connect initial sync (useDeviceAutoSync) passes a
-  // `{ favorites: true, keyboard }` scope. typing-analytics is pulled
-  // separately when the Analyze panel opens — skip it here so the
-  // connect progress bar stays short. Run logs are excluded the same
-  // way (no per-uid keystroke count to show on the connect progress
-  // bar either) — see .claude/rules/settings-persistence.md Row E.
-  if (typeof scope === 'object' && 'favorites' in scope && (isAnalyticsSyncUnit(syncUnit) || isRunLogSyncUnit(syncUnit))) {
-    return false
-  }
-  // Lazy: when scope is 'all' only download keyboards/<uid>/* that already exist locally.
-  // Explicit keyboard scopes always download in full.
-  if (syncUnit.startsWith('keyboards/') && scope === 'all') {
-    const uid = syncUnit.split('/')[1]
-    return !!uid && localKeyboardUids.has(uid)
-  }
-  return true
-}
-
-async function requireSyncCredentials(): Promise<SyncCredentialResult> {
-  const authStatus = await getAuthStatus()
-  if (!authStatus.authenticated) return { ok: false, reason: 'unauthenticated' }
-  return retrievePasswordResult()
-}
-
-// --- Remote data inspection ---
-
-async function fetchValidatedDataFiles(): Promise<{ password: string; dataFiles: DriveFile[] } | null> {
-  const credentials = await requireSyncCredentials()
-  if (!credentials.ok) return null
-  const { password } = credentials
-  const remoteFiles = await listFiles()
-
-  await validatePasswordCheck(password, remoteFiles)
-
-  const passwordCheckFileName = driveFileName(PASSWORD_CHECK_UNIT)
-  const dataFiles = remoteFiles.filter((f) => f.name !== passwordCheckFileName)
-  return { password, dataFiles }
-}
-
-async function findUndecryptableFiles(password: string, dataFiles: DriveFile[]): Promise<UndecryptableFile[]> {
-  const undecryptable: UndecryptableFile[] = []
-  const limit = pLimit(SYNC_CONCURRENCY)
-  await Promise.allSettled(
-    dataFiles.map((file) =>
-      limit(async () => {
-        try {
-          const envelope = await downloadFile(file.id)
-          await decrypt(envelope, password)
-        } catch {
-          undecryptable.push({
-            fileId: file.id,
-            fileName: file.name,
-            syncUnit: syncUnitFromFileName(file.name),
-          })
-        }
-      }),
-    ),
-  )
-  return undecryptable
-}
-
-export async function listUndecryptableFiles(): Promise<UndecryptableFile[]> {
-  const result = await fetchValidatedDataFiles()
-  if (!result) return []
-  return findUndecryptableFiles(result.password, result.dataFiles)
-}
-
-/** Shared by the no-credentials early return and (spread into) the
- *  success path below, so a future new `SyncDataScanResult` field only
- *  needs a default added here once instead of in both literals. */
-const EMPTY_SCAN_RESULT: SyncDataScanResult = {
-  keyboards: [],
-  keyboardNames: {},
-  favorites: [],
-  i18nPacks: [],
-  themePacks: [],
-  keyLabels: false,
-  typingTestTexts: false,
-  hasI18nData: false,
-  hasThemesData: false,
-  undecryptable: [],
-}
-
-export async function scanRemoteData(): Promise<SyncDataScanResult> {
-  const result = await fetchValidatedDataFiles()
-  if (!result) {
-    return EMPTY_SCAN_RESULT
-  }
-  const { password, dataFiles } = result
-
-  // Categorize from filenames (no download needed)
-  const keyboardUids = new Set<string>()
-  const favoriteTypes = new Set<string>()
-  const i18nPackIds = new Set<string>()
-  const themePackIds = new Set<string>()
-  let keyLabelsFound = false
-  let typingTestTextsFound = false
-  let i18nIndexFound = false
-  let themesIndexFound = false
-  for (const file of dataFiles) {
-    const syncUnit = syncUnitFromFileName(file.name)
-    if (!syncUnit) continue
-    if (syncUnit.startsWith('keyboards/')) {
-      const uid = syncUnit.split('/')[1]
-      if (uid) keyboardUids.add(uid)
-    } else if (syncUnit.startsWith('favorites/')) {
-      const type = syncUnit.split('/')[1]
-      if (type) favoriteTypes.add(type)
-    } else if (syncUnit.startsWith('i18n/packs/')) {
-      const packId = syncUnit.slice('i18n/packs/'.length)
-      if (packId) i18nPackIds.add(packId)
-    } else if (syncUnit.startsWith('themes/packs/')) {
-      const packId = syncUnit.slice('themes/packs/'.length)
-      if (packId) themePackIds.add(packId)
-    } else if (syncUnit === KEY_LABEL_SYNC_UNIT) {
-      keyLabelsFound = true
-    } else if (syncUnit === TYPING_TEST_TEXT_SYNC_UNIT) {
-      typingTestTextsFound = true
-    } else if (syncUnit === I18N_INDEX_SYNC_UNIT) {
-      // The index can outlive every pack id it once listed (all
-      // tombstoned and GC'd) — tracked separately so `hasI18nData`
-      // below still surfaces a resettable target in that 30-day dead
-      // zone instead of going silent just because `i18nPackIds` is empty.
-      i18nIndexFound = true
-    } else if (syncUnit === THEME_INDEX_SYNC_UNIT) {
-      themesIndexFound = true
-    }
-  }
-
-  // Use whatever names are already in the local meta index (populated by executeSync/backfill
-  // and the LIST_STORED_KEYBOARDS safety net). scanRemoteData stays read-only here so it
-  // doesn't trigger extra downloads or writes.
-  const metaIndex = await readKeyboardMetaIndex()
-  const metaMap = getActiveKeyboardMetaMap(metaIndex)
-  const keyboardNames: Record<string, string> = {}
-  for (const uid of keyboardUids) {
-    const name = metaMap.get(uid)
-    if (name) keyboardNames[uid] = name
-  }
-
-  const undecryptable = await findUndecryptableFiles(password, dataFiles)
-
-  return {
-    ...EMPTY_SCAN_RESULT,
-    keyboards: [...keyboardUids],
-    keyboardNames,
-    favorites: [...favoriteTypes],
-    i18nPacks: [...i18nPackIds],
-    themePacks: [...themePackIds],
-    keyLabels: keyLabelsFound,
-    typingTestTexts: typingTestTextsFound,
-    hasI18nData: i18nIndexFound || i18nPackIds.size > 0,
-    hasThemesData: themesIndexFound || themePackIds.size > 0,
-    undecryptable,
-  }
-}
-
-/** Download and decrypt a remote sync unit bundle without merging into local. */
-export async function fetchRemoteBundle(syncUnit: string): Promise<SyncBundle | null> {
-  const result = await fetchValidatedDataFiles()
-  if (!result) return null
-  const { password, dataFiles } = result
-  const targetName = driveFileName(syncUnit)
-  const file = dataFiles.find((f) => f.name === targetName)
-  if (!file) return null
-  try {
-    const envelope = await downloadFile(file.id)
-    const plaintext = await decrypt(envelope, password)
-    return JSON.parse(plaintext) as SyncBundle
-  } catch {
-    return null
-  }
-}
-
-// --- Non-destructive password change ---
-
-export async function changePassword(newPassword: string): Promise<void> {
-  if (isSyncing) throw new Error('sync.changePasswordInProgress')
-  isSyncing = true
-  try {
-    const credentials = await requireSyncCredentials()
-    if (!credentials.ok) throw new SyncCredentialError(credentials.reason)
-    const oldPassword = credentials.password
-    if (newPassword === oldPassword) throw new Error('sync.samePassword')
-    const remoteFiles = await listFiles()
-
-    // Validate old password against password-check first
-    await validatePasswordCheck(oldPassword, remoteFiles)
-
-    const passwordCheckFileName = driveFileName(PASSWORD_CHECK_UNIT)
-    const dataFiles = remoteFiles.filter((f) => f.name !== passwordCheckFileName)
-
-    // Phase 1: Download + decrypt all files (fail-fast on any error)
-    const limit = pLimit(SYNC_CONCURRENCY)
-    const decrypted = await Promise.all(
-      dataFiles.map((file) =>
-        limit(async () => {
-          const envelope = await downloadFile(file.id)
-          try {
-            const plaintext = await decrypt(envelope, oldPassword)
-            return { file, plaintext, syncUnit: envelope.syncUnit }
-          } catch {
-            throw new Error('sync.changePasswordUndecryptable')
-          }
-        }),
-      ),
-    )
-
-    // Phase 2: Re-encrypt + upload with new password (overwrite)
-    await Promise.all(
-      decrypted.map(({ file, plaintext, syncUnit }) =>
-        limit(async () => {
-          const newEnvelope = await encrypt(plaintext, newPassword, syncUnit)
-          await uploadFile(file.name, newEnvelope, file.id)
-        }),
-      ),
-    )
-
-    // Phase 3: Recreate password-check with new password
-    const existingPc = remoteFiles.find((f) => f.name === passwordCheckFileName)
-    const pcEnvelope = await encrypt(PASSWORD_CHECK_PAYLOAD, newPassword, PASSWORD_CHECK_UNIT)
-    await uploadFile(passwordCheckFileName, pcEnvelope, existingPc?.id)
-
-    await storePassword(newPassword)
-    resetPasswordCheckCache()
-  } finally {
-    isSyncing = false
-  }
-}
+export { isAnalyticsSyncUnit, isRunLogSyncUnit } from './sync-bundle'
 
 // Re-export bundle functions for backward compatibility
 export { readIndexFile, bundleSyncUnit, collectAllSyncUnits } from './sync-bundle'
 
-// --- Password check validation ---
-
-class PasswordMismatchError extends Error {
-  constructor() {
-    super('sync.passwordMismatch')
-    this.name = 'PasswordMismatchError'
-  }
-}
-
-async function validatePasswordCheck(
-  password: string,
-  remoteFiles: DriveFile[],
-): Promise<void> {
-  const fileName = driveFileName(PASSWORD_CHECK_UNIT)
-  const existing = remoteFiles.find((f) => f.name === fileName)
-
-  if (existing) {
-    const envelope = await downloadFile(existing.id)
-    try {
-      await decrypt(envelope, password)
-    } catch {
-      throw new PasswordMismatchError()
-    }
-  } else {
-    const envelope = await encrypt(PASSWORD_CHECK_PAYLOAD, password, PASSWORD_CHECK_UNIT)
-    await uploadFile(fileName, envelope)
-  }
-  passwordCheckValidated = true
-}
-
-export function resetPasswordCheckCache(): void {
-  passwordCheckValidated = false
-}
-
-export async function checkPasswordCheckExists(): Promise<boolean> {
-  const remoteFiles = await listFiles()
-  const fileName = driveFileName(PASSWORD_CHECK_UNIT)
-  return remoteFiles.some((f) => f.name === fileName)
-}
-
-export async function setPasswordAndValidate(password: string): Promise<void> {
-  await storePassword(password)
-  resetPasswordCheckCache()
-  try {
-    const remoteFiles = await listFiles()
-    await validatePasswordCheck(password, remoteFiles)
-  } catch (err) {
-    await clearPassword()
-    throw err
-  }
-}
-
-// --- Sync operations ---
-
-async function uploadSyncUnit(
-  syncUnit: string,
-  password: string,
-  remoteFiles?: DriveFile[],
-): Promise<void> {
-  // i18n/theme pack bodies: snapshot the local file's mtime BEFORE
-  // bundling — bundling/encrypting/uploading all happen without holding
-  // the store's write lock, so a user save (savePack/renamePack) can
-  // land in that window. This snapshot is the CAS baseline
-  // pinPackBodyMtimeAfterUpload needs below to detect that race — see
-  // its doc for why a blind post-upload pin would otherwise stamp a
-  // fresher local edit with this upload's stale Drive time.
-  const packBodyRef = parsePackBodySyncUnit(syncUnit)
-  const packBodyMtimeSnapshot = packBodyRef ? await statPackBodyLocalMtime(packBodyRef) : null
-
-  const bundle = await bundleSyncUnit(syncUnit)
-  if (!bundle) return
-
-  const plaintext = JSON.stringify(bundle)
-  const envelope = await encrypt(plaintext, password, syncUnit)
-
-  const files = remoteFiles ?? await listFiles()
-  const targetName = driveFileName(syncUnit)
-  const existing = files.find((f) => f.name === targetName)
-
-  const uploaded = await uploadFile(targetName, envelope, existing?.id)
-
-  // Post-upload bookkeeping: record a successful cloud upload for
-  // per-day units so the reconcile logic can later distinguish
-  // "never uploaded" from "uploaded then remotely deleted".
-  const dayRef = parseTypingAnalyticsDeviceDaySyncUnit(syncUnit)
-  if (dayRef) await recordDayUploaded(dayRef)
-
-  // i18n/theme pack bodies: a local-wins upload just gave Drive a fresh
-  // `modifiedTime` — pin the local file's mtime to it (rather than
-  // leaving it at "now") so a locally-ahead wall clock can't make this
-  // file look newer than Drive's own stamped time forever. See
-  // pinPackBodyMtimeAfterUpload's doc for the full clock-skew rationale
-  // and why the snapshot above is passed through as a CAS guard.
-  if (packBodyRef) await pinPackBodyMtimeAfterUpload(packBodyRef, uploaded.modifiedTime, packBodyMtimeSnapshot)
-}
-
-/** Add `{uid}|{hash}` → utcDay to sync-state.uploaded after a
- * successful cloud upload. Idempotent: the list is kept sorted and
- * duplicate-free so repeated uploads of the current-day file don't
- * grow the array. */
-async function recordDayUploaded(dayRef: {
-  uid: string
-  machineHash: string
-  utcDay: UtcDay
-}): Promise<void> {
-  const userData = app.getPath('userData')
-  const ownHash = await getMachineHash()
-  const state = (await loadSyncState(userData)) ?? emptySyncState(ownHash)
-  const pointerKey = readPointerKey(dayRef.uid, dayRef.machineHash)
-  const existing = new Set(state.uploaded[pointerKey] ?? [])
-  if (existing.has(dayRef.utcDay)) return
-  existing.add(dayRef.utcDay)
-  state.uploaded[pointerKey] = Array.from(existing).sort()
-  state.last_synced_at = Date.now()
-  await saveSyncState(userData, state)
-}
-
-/** Write a downloaded per-day JSONL under the owning device's `{hash}/`
- * directory and apply every row in the file. Each day is a distinct
- * file so a partial download of one day does not affect other days for
- * the same remote hash. No-op when the unit's machineHash matches our
- * own. */
-async function mergeDeviceDayBundle(
-  remoteBundle: SyncBundle,
-  dayRef: { uid: string; machineHash: string; utcDay: UtcDay },
-  userData: string,
-  ownHash: string,
-): Promise<void> {
-  if (dayRef.machineHash === ownHash) return
-  const data = remoteBundle.files['data.jsonl']
-  if (!data) return
-
-  const localPath = deviceDayJsonlPath(userData, dayRef.uid, dayRef.machineHash, dayRef.utcDay)
-  await mkdir(deviceDayDir(userData, dayRef.uid, dayRef.machineHash), { recursive: true })
-  await writeFile(localPath, data, 'utf-8')
-
-  // Per-day bundles are replayed in full. The LWW merge is idempotent,
-  // so re-applying every row in the file is cheap, correct, and avoids
-  // any per-hash `afterId` bookkeeping at the merge layer.
-  const { rows } = await readRows(localPath)
-  if (rows.length > 0) {
-    applyRowsToCache(getTypingAnalyticsDB(), rows)
-  }
-  const state = (await loadSyncState(userData)) ?? emptySyncState(ownHash)
-  state.last_synced_at = Date.now()
-  await saveSyncState(userData, state)
-}
-
-// Merges remote bundle into local state, returns whether remote needs update
-async function mergeSyncUnit(
-  syncUnit: string,
-  envelope: SyncEnvelope,
-  password: string,
-  remoteModifiedTime: string,
-): Promise<boolean> {
-  const plaintext = await decrypt(envelope, password)
-  const remoteBundle = JSON.parse(plaintext) as SyncBundle
-
-  // Handle meta/keyboard-names (entry-level LWW, no data files)
-  if (syncUnit === KEYBOARD_META_SYNC_UNIT) {
-    const remoteIndex = remoteBundle.index as KeyboardMetaIndex
-    const { remoteNeedsUpdate } = await applyRemoteKeyboardMetaIndex(remoteIndex)
-    return remoteNeedsUpdate
-  }
-
-  const parts = syncUnit.split('/')
-  const userData = app.getPath('userData')
-
-  // `syncUnit` here is derived from `syncUnitFromFileName(file.name)` —
-  // a Drive filename is attacker-reachable data (anyone who can write
-  // to this appData folder), and a crafted name (e.g.
-  // `favorites_../../evil.enc` or `keyboards_../../evil_settings.enc`)
-  // can make a capture group in that regex contain a path separator.
-  // Every `/`-split segment must pass `isSafePathSegment` before ANY
-  // branch below joins it into a filesystem path (the settings branch
-  // and the generic index-based tail both do) — a single unsafe segment
-  // throws the same `MalformedSyncBundleError` a corrupt bundle shape
-  // does, so the sync poll's unchanged-revision skip applies here too
-  // instead of retrying a permanently-hostile filename every 3 minutes.
-  if (parts.some((part) => !isSafePathSegment(part))) {
-    throw new MalformedSyncBundleError(syncUnit)
-  }
-
-  // Typing-analytics JSONL: each file is owned by one device. Skip our
-  // own hash so a stale remote never clobbers freshly-flushed local
-  // rows. For a remote device's file we overwrite the local copy and
-  // replay only the newly-appended rows into the cache.
-  const dayRef = parseTypingAnalyticsDeviceDaySyncUnit(syncUnit)
-  if (dayRef) {
-    await mergeDeviceDayBundle(remoteBundle, dayRef, userData, await getMachineHash())
-    return false
-  }
-
-  // Handle settings sync unit (single-file LWW)
-  if (parts.length === 3 && parts[0] === 'keyboards' && parts[2] === 'settings') {
-    const dir = join(userData, 'sync', 'keyboards', parts[1])
-    await mkdir(dir, { recursive: true })
-
-    const filePath = join(dir, 'pipette_settings.json')
-    const remoteContent = remoteBundle.files['pipette_settings.json']
-    if (!remoteContent) return false
-
-    let localTime = 0
-    try {
-      const raw = await readFile(filePath, 'utf-8')
-      const local = JSON.parse(raw) as { _updatedAt?: string }
-      localTime = safeTimestamp(local._updatedAt)
-    } catch { /* no local settings */ }
-
-    const remoteSettings = JSON.parse(remoteContent) as { _updatedAt?: string }
-    const remoteTime = safeTimestamp(remoteSettings._updatedAt)
-
-    if (remoteTime > localTime) {
-      await writeFile(filePath, remoteContent, 'utf-8')
-      return false
-    }
-    return localTime > remoteTime
-  }
-
-  // Handle "i18n/index" / "themes/index" — the language/theme pack
-  // roster, entry-level LWW (same mergeEntries/gcTombstones machinery
-  // as favorites/key-labels/etc. below, applied to the pack meta shape
-  // instead of a whole-file "newer roster wins wholesale" comparison —
-  // see mergePackIndexBundle's doc for why the old file-level strategy
-  // was a data-loss bug).
-  if (syncUnit === I18N_INDEX_SYNC_UNIT || syncUnit === THEME_INDEX_SYNC_UNIT) {
-    return mergePackIndexBundle(syncUnit, remoteBundle)
-  }
-
-  // Handle "i18n/packs/{packId}" / "themes/packs/{packId}" — a single
-  // pack body, file-level LWW using the Drive file's own modifiedTime
-  // (the bundle carries no timestamp of its own — see
-  // mergePackBodyBundle's doc).
-  const packBodyRef = parsePackBodySyncUnit(syncUnit)
-  if (packBodyRef) {
-    return mergePackBodyBundle(packBodyRef, remoteBundle, remoteModifiedTime)
-  }
-
-  // Handle index-based sync units (favorites, snapshots, analyze-filter,
-  // key-label, typing-test-text, run-log)
-  const basePath = join(userData, 'sync', ...parts)
-  await mkdir(basePath, { recursive: true })
-
-  const localIndex = await readIndexFile(basePath)
-  // Both sides' index shape is a union of each possible sync unit's own
-  // index type, which a generic function call can't unify against — every
-  // constituent reached on this branch is an EntryMeta[] at runtime.
-  // i18n/theme/keyboard-meta bundles are intercepted by the dedicated
-  // branches above and never reach here.
-  //
-  // Deliberate coerce-vs-throw asymmetry below: a non-array `.entries` on
-  // the LOCAL side is coerced to `[]` (trusted data — a missing/corrupt
-  // local index is just an empty starting point, not a reason to fail
-  // the merge), while the same shape on the REMOTE side throws
-  // MalformedSyncBundleError (untrusted, attacker-reachable data — see
-  // its doc for why silently coercing there would be the wrong call).
-  const localEntries = gcTombstones(
-    Array.isArray(localIndex?.entries) ? (localIndex.entries as EntryMeta[]) : [],
-  )
-  // A remote bundle is attacker-reachable data (anyone who can write to
-  // this sync unit's Drive file) — validate `.entries` is actually an
-  // array before handing it to gcTombstones/mergeEntries instead of
-  // letting a malformed shape throw an opaque TypeError deep inside
-  // them. See MalformedSyncBundleError's doc for how callers contain
-  // this per-unit.
-  const remoteIndexEntries = (remoteBundle.index as { entries?: unknown } | undefined)?.entries
-  if (!Array.isArray(remoteIndexEntries)) {
-    throw new MalformedSyncBundleError(syncUnit)
-  }
-  const remoteEntries = gcTombstones(remoteIndexEntries as EntryMeta[])
-
-  // Merge entries (both sides GC'd to prevent expired-tombstone upload loops).
-  // Run-log retention is a deterministic trim (not reject-at-cap like
-  // analyze-filter/snapshot), applied as part of this same merge via
-  // `runLogRetentionMax` — see applyRunLogRetention's doc comment for why
-  // ranking by immutable `startedAt` lets two devices that independently
-  // exceeded the cap converge on the same kept set.
-  const preserveLocalOrder = syncUnit === KEY_LABEL_SYNC_UNIT
-  const runLogRetentionMax = isRunLogSyncUnit(syncUnit) ? MAX_RUN_LOGS_PER_KEYBOARD : undefined
-  const result = mergeEntries(localEntries, remoteEntries, { preserveLocalOrder, runLogRetentionMax })
-
-  // Copy files from remote bundle for entries that remote won. Every
-  // remote entry's filename must pass isSafePathSegment before it's
-  // joined into a local path — a remote bundle is attacker-reachable
-  // data (anyone who can write to this sync unit's Drive file), and this
-  // is the generic write site every index-based sync unit (favorites,
-  // snapshots, analyze-filter, key-label, typing-test-text, run logs)
-  // funnels through, so the guard has to live here rather than per-unit.
-  let unsafeRemoteFilenames = 0
-  for (const filename of result.remoteFilesToCopy) {
-    if (!isSafePathSegment(filename)) {
-      unsafeRemoteFilenames++
-      continue
-    }
-    if (filename in remoteBundle.files) {
-      await writeFile(join(basePath, filename), remoteBundle.files[filename], 'utf-8')
-    }
-  }
-  if (unsafeRemoteFilenames > 0) {
-    log('warn', `sync: skipped ${unsafeRemoteFilenames} unsafe remote filename(s) for ${syncUnit}`)
-  }
-
-  // Write merged index
-  let mergedIndex = localIndex
-    ? { ...localIndex, entries: result.entries }
-    : remoteBundle.index
-
-  // For the remote-only-so-far branch (no local index yet) the retention
-  // trim above must still land in what gets written — narrow override
-  // kept to exactly this sync unit.
-  if (isRunLogSyncUnit(syncUnit)) {
-    mergedIndex = { ...mergedIndex, entries: result.entries }
-  }
-
-  await writeFile(
-    join(basePath, 'index.json'),
-    JSON.stringify(mergedIndex, null, 2),
-    'utf-8',
-  )
-
-  // Unlink files for entries retention evicted during the merge above —
-  // best-effort, a file already gone is not an error. Always empty for
-  // every sync unit except run logs (runLogRetentionMax above). Inlined
-  // here (rather than shared with typing-run-log-store.ts's own local-save
-  // eviction) since this module already owns basePath/fs for this branch.
-  for (const meta of result.evicted) {
-    if (!isSafePathSegment(meta.filename)) continue
-    try {
-      await unlink(join(basePath, meta.filename))
-    } catch {
-      // best-effort
-    }
-  }
-
-  return result.remoteNeedsUpdate
-}
-
-// Merges with remote, uploads if local has changes remote doesn't have.
-// Takes the full DriveFile (not just its id) because it's the single
-// choke point every merge path funnels through (download sync, polling,
-// analytics sync, upload-time merge-before-upload) — an i18n/theme
-// pack-body unit can skip the download+decrypt entirely here when local
-// is already known to be newer (see `packBodyLocalWins`'s doc).
-async function mergeWithRemote(
-  remoteFile: DriveFile,
-  syncUnit: string,
-  password: string,
-  remoteFiles?: DriveFile[],
-): Promise<void> {
-  const packBodyRef = parsePackBodySyncUnit(syncUnit)
-  if (packBodyRef && await packBodyLocalWins(packBodyRef, remoteFile.modifiedTime)) {
-    await uploadSyncUnit(syncUnit, password, remoteFiles)
-    return
-  }
-
-  const envelope = await downloadFile(remoteFile.id)
-  const needsUpload = await mergeSyncUnit(syncUnit, envelope, password, remoteFile.modifiedTime)
-  if (needsUpload) {
-    await uploadSyncUnit(syncUnit, password, remoteFiles)
-  }
-}
-
-async function syncOrUpload(
-  syncUnit: string,
-  password: string,
-  remoteFiles: DriveFile[],
-): Promise<void> {
-  const targetName = driveFileName(syncUnit)
-  const remoteFile = remoteFiles.find((f) => f.name === targetName)
-
-  if (remoteFile) {
-    await mergeWithRemote(remoteFile, syncUnit, password, remoteFiles)
-  } else {
-    await uploadSyncUnit(syncUnit, password, remoteFiles)
-  }
-}
-
-/** Real outcome of an `executeSync` call — distinct from the `void`
- *  return the caller used to get, which made a busy-race skip and a
- *  missing-credentials skip both look identical to a fully-completed
- *  sync (neither throws; both just emit progress and return). Threaded
- *  through SYNC_EXECUTE's IPC result as `status`/`skipReason` — see
- *  `SyncOperationResult`'s doc in shared/types/sync.ts for why `success`
- *  itself is deliberately left alone. */
-export interface SyncExecuteResult {
-  status: SyncExecuteStatus
-  /** Populated only when `status === 'skipped'`. */
-  skipReason?: SyncSkipReason
-  /** Populated only when `status === 'partial'`. */
-  failedUnits?: string[]
-}
-
-export async function executeSync(
-  direction: 'download' | 'upload',
-  scope: SyncScope = 'all',
-): Promise<SyncExecuteResult> {
-  if (isSyncing) return { status: 'skipped', skipReason: 'busy' }
-  isSyncing = true
-
-  try {
-    const credentials = await requireSyncCredentials()
-    if (!credentials.ok) {
-      emitProgress({
-        direction,
-        status: 'error',
-        reason: credentials.reason,
-        message: syncCredentialI18nKey('readiness', credentials.reason),
-      })
-      return { status: 'skipped', skipReason: credentials.reason }
-    }
-    const password = credentials.password
-
-    emitProgress({ direction, status: 'syncing', message: 'Starting sync...' })
-
-    const initialFiles = await listFiles()
-
-    // Force password re-validation on scope 'all' (changePassword, listUndecryptable)
-    // Scoped syncs (including manual sync) respect the cached validation;
-    // decryption errors during actual file processing serve as implicit validation
-    if (scope === 'all' || !passwordCheckValidated) {
-      await validatePasswordCheck(password, initialFiles)
-    }
-
-    let failedUnits: string[]
-    if (direction === 'download') {
-      failedUnits = await executeDownloadSync(password, initialFiles, scope)
-      if (scope === 'all') {
-        const { resolved } = await backfillKeyboardMeta(password, initialFiles)
-        if (resolved > 0) {
-          pendingChanges.add(KEYBOARD_META_SYNC_UNIT)
-          broadcastPendingStatus()
-        }
-      }
-    } else {
-      failedUnits = await executeUploadSync(password, initialFiles, scope)
-      // Clear pending changes matching the scope, then re-add failed units
-      for (const unit of pendingChanges) {
-        if (matchesScope(unit, scope)) pendingChanges.delete(unit)
-      }
-      for (const unit of failedUnits) {
-        pendingChanges.add(unit)
-      }
-      broadcastPendingStatus()
-    }
-
-    if (failedUnits.length === 0) {
-      emitProgress({ direction, status: 'success', message: 'Sync complete' })
-      return { status: 'completed' }
-    } else {
-      emitProgress({
-        direction,
-        status: 'partial',
-        message: `${failedUnits.length} sync unit(s) failed`,
-        failedUnits,
-      })
-      return { status: 'partial', failedUnits }
-    }
-  } catch (err) {
-    emitProgress({
-      direction,
-      status: 'error',
-      message: errorMessage(err, 'Sync failed'),
-    })
-    throw err
-  } finally {
-    isSyncing = false
-  }
-}
-
-async function executeDownloadSync(
-  password: string,
-  prefetchedFiles?: DriveFile[],
-  scope: SyncScope = 'all',
-): Promise<string[]> {
-  const remoteFiles = prefetchedFiles ?? await listFiles()
-  updateRemoteState(remoteFiles) // Always record full remote state for polling
-
-  const localKeyboardUids = await listLocalKeyboardUids()
-  // {file, syncUnit} pairs resolved once here rather than re-parsing the
-  // filename again inside the download loop below.
-  const filesToDownload = remoteFiles.flatMap((file) => {
-    const syncUnit = syncUnitFromFileName(file.name)
-    if (!syncUnit || !shouldDownloadSyncUnit(syncUnit, scope, localKeyboardUids)) return []
-    return [{ file, syncUnit }]
-  })
-
-  const total = filesToDownload.length
-  let completed = 0
-  const failedUnits: string[] = []
-  const limit = pLimit(SYNC_CONCURRENCY)
-
-  await Promise.allSettled(
-    filesToDownload.map(({ file: remoteFile, syncUnit }) =>
-      limit(async () => {
-        completed++
-
-        emitProgress({
-          direction: 'download',
-          status: 'syncing',
-          syncUnit,
-          current: completed,
-          total,
-        })
-
-        try {
-          await mergeWithRemote(remoteFile, syncUnit, password, remoteFiles)
-        } catch (err) {
-          failedUnits.push(syncUnit)
-          emitProgress({
-            direction: 'download',
-            status: 'error',
-            syncUnit,
-            message: errorMessage(err, 'Download failed'),
-          })
-        }
-      }),
-    ),
-  )
-
-  // Pass-level GC — see pack-gc.ts's doc for why this must never be
-  // triggered from inside a single unit's own merge callback above, and
-  // for why `failedUnits` is passed separately from the full attempted
-  // list (a failed unit skips that store's sweep, not the whole GC call).
-  await runPackGcAfterPass(filesToDownload.map((f) => f.syncUnit), failedUnits)
-
-  return failedUnits
-}
-
-/** Scan the remote file list for per-day typing-analytics units owned
- * by `ownHash`, grouped by keyboard uid. Units with a malformed
- * filename are skipped. */
-function collectRemoteOwnHashDays(
-  remoteFiles: DriveFile[],
-  ownHash: string,
-): Map<string, Map<UtcDay, DriveFile>> {
-  const perUid = new Map<string, Map<UtcDay, DriveFile>>()
-  for (const file of remoteFiles) {
-    const unit = syncUnitFromFileName(file.name)
-    if (!unit) continue
-    const ref = parseTypingAnalyticsDeviceDaySyncUnit(unit)
-    if (!ref || ref.machineHash !== ownHash) continue
-    let byDay = perUid.get(ref.uid)
-    if (!byDay) {
-      byDay = new Map<UtcDay, DriveFile>()
-      perUid.set(ref.uid, byDay)
-    }
-    byDay.set(ref.utcDay, file)
-  }
-  return perUid
-}
-
-/** Reconcile the own-hash cloud state with local + uploaded bookkeeping
- * before the regular upload pass runs. Three transitions are applied:
- *
- *   Rule 2 — `uploaded` has day X, local does not: user or Local-delete
- *     removed the file locally → delete the cloud copy as well.
- *   Rule 3 — `uploaded` has day X, cloud does not: a Sync-delete from
- *     another device or a GC step removed the cloud copy → drop the
- *     local file and let the next cache rebuild resync (rows added
- *     post-delete are preserved because they were never in `uploaded`).
- *   Orphan — when `reconciled_at` is pending for (uid, ownHash), also
- *     delete any cloud day that is neither in local nor in `uploaded`
- *     (leftover from a previous install / pre-migration state).
- *
- * Rules 2 and 3 run on every pass; orphan cleanup only on the first
- * pass after a cache rebuild or fresh install, then `reconciled_at`
- * is timestamped so the expensive listing is skipped afterwards. */
-async function reconcileOwnHashTypingAnalytics(
-  remoteFiles: DriveFile[],
-  userData: string,
-  ownHash: string,
-): Promise<{ state: TypingSyncState; mutated: boolean }> {
-  const state = (await loadSyncState(userData)) ?? emptySyncState(ownHash)
-  const remotePerUid = collectRemoteOwnHashDays(remoteFiles, ownHash)
-
-  // Every uid that appears in any of the three sources needs a pass:
-  // local files, uploaded bookkeeping, or remote cloud listing. Union
-  // them so a fully-remote-only uid (no local files left) still gets
-  // reconciled.
-  const candidateUids = new Set<string>()
-  for (const key of Object.keys(state.uploaded)) {
-    const parts = key.split('|')
-    if (parts.length === 2 && parts[1] === ownHash) candidateUids.add(parts[0])
-  }
-  for (const uid of remotePerUid.keys()) candidateUids.add(uid)
-  try {
-    for (const entry of await readdir(join(userData, 'sync', 'keyboards'), { withFileTypes: true })) {
-      if (entry.isDirectory()) candidateUids.add(entry.name)
-    }
-  } catch { /* no keyboards dir */ }
-
-  let mutated = false
-  for (const uid of candidateUids) {
-    const pointerKey = readPointerKey(uid, ownHash)
-    const localDays = new Set<UtcDay>(await listDeviceDays(userData, uid, ownHash))
-    const uploadedDays = new Set<UtcDay>(state.uploaded[pointerKey] ?? [])
-    const cloudDays = remotePerUid.get(uid) ?? new Map<UtcDay, DriveFile>()
-
-    // Rule 2: uploaded but not local — delete from cloud.
-    for (const day of Array.from(uploadedDays)) {
-      if (localDays.has(day)) continue
-      const cloudFile = cloudDays.get(day)
-      if (cloudFile) {
-        try {
-          await deleteFile(cloudFile.id)
-        } catch (err) {
-          log('warn', `typing-analytics cloud delete failed for ${uid} ${day}: ${String(err)}`)
-        }
-        cloudDays.delete(day)
-      }
-      uploadedDays.delete(day)
-      mutated = true
-    }
-
-    // Rule 3: uploaded but not cloud — another device Sync-deleted us.
-    for (const day of Array.from(uploadedDays)) {
-      if (cloudDays.has(day)) continue
-      try {
-        await unlink(deviceDayJsonlPath(userData, uid, ownHash, day))
-      } catch (err) {
-        if ((err as NodeJS.ErrnoException).code !== 'ENOENT') {
-          log('warn', `typing-analytics local delete failed for ${uid} ${day}: ${String(err)}`)
-        }
-      }
-      localDays.delete(day)
-      uploadedDays.delete(day)
-      mutated = true
-    }
-
-    // Orphan cleanup on first reconcile only. Cloud days that are
-    // neither local nor in `uploaded` are leftovers (pre-migration
-    // flat bundles converted to per-day, or data from a removed
-    // install). Deleting them avoids surprising re-download prompts.
-    if (isReconcilePending(state, uid, ownHash)) {
-      for (const [day, cloudFile] of Array.from(cloudDays.entries())) {
-        if (localDays.has(day) || uploadedDays.has(day)) continue
-        try {
-          await deleteFile(cloudFile.id)
-        } catch (err) {
-          log('warn', `typing-analytics orphan delete failed for ${uid} ${day}: ${String(err)}`)
-        }
-        cloudDays.delete(day)
-      }
-      state.reconciled_at[pointerKey] = Date.now()
-      mutated = true
-    }
-
-    // Persist the trimmed uploaded list (sorted for determinism so
-    // the JSON-on-disk diff stays stable).
-    state.uploaded[pointerKey] = Array.from(uploadedDays).sort()
-  }
-
-  if (mutated) {
-    state.last_synced_at = Date.now()
-    await saveSyncState(userData, state)
-  }
-  return { state, mutated }
-}
-
-/** Snapshot of the user's appData Drive listing as a name-only set,
- * for callers that need many existence checks (e.g. import). Returns
- * `null` when the user is unauthenticated so the caller can fall back
- * to a local-only check rather than rejecting outright. */
-export async function listRemoteFileNames(): Promise<Set<string> | null> {
-  const credentials = await requireSyncCredentials()
-  if (!credentials.ok) return null
-  const remoteFiles = await listFiles()
-  return new Set(remoteFiles.map((f) => f.name))
-}
-
-/** True iff cloud currently holds at least one typing per-day file
- * owned by a non-own device. Used to decide whether the Sync > Typing
- * nav subtree is worth showing at all — a single listing is much
- * cheaper than expanding every keyboard. Returns `false` when the
- * user is unauthenticated. */
-export async function hasAnyRemoteTypingData(): Promise<boolean> {
-  const credentials = await requireSyncCredentials()
-  if (!credentials.ok) return false
-  const ownHash = await getMachineHash()
-  const remoteFiles = await listFiles()
-  for (const file of remoteFiles) {
-    const unit = syncUnitFromFileName(file.name)
-    if (!unit) continue
-    const ref = parseTypingAnalyticsDeviceDaySyncUnit(unit)
-    if (!ref || ref.machineHash === ownHash) continue
-    return true
-  }
-  return false
-}
-
-/** Distinct remote machineHash values (non-own) that cloud currently
- * holds any per-day file for under `uid`. Used by the Sync > Typing
- * subtree to discover remote devices before the user has ever opened
- * one — the cache-only `listRemoteHashesForUid` misses hashes that
- * haven't been merged locally yet. Sorted for stable UI order. */
-export async function listRemoteTypingHashesForUidFromCloud(
-  uid: string,
-): Promise<string[]> {
-  const credentials = await requireSyncCredentials()
-  if (!credentials.ok) return []
-  const ownHash = await getMachineHash()
-  const remoteFiles = await listFiles()
-  const hashes = new Set<string>()
-  for (const file of remoteFiles) {
-    const unit = syncUnitFromFileName(file.name)
-    if (!unit) continue
-    const ref = parseTypingAnalyticsDeviceDaySyncUnit(unit)
-    if (!ref || ref.uid !== uid || ref.machineHash === ownHash) continue
-    hashes.add(ref.machineHash)
-  }
-  return Array.from(hashes).sort()
-}
-
-/** List the UTC days that cloud currently holds for a remote device
- * `(uid, machineHash)`. Returned in ascending lexicographic order so
- * callers can feed the list straight into a Sync > Typing > Device
- * tree without post-processing. An unauthenticated / network-failed
- * call returns an empty array — UIs surface the network error via
- * scanRemoteData or the sync progress channel separately. */
-export async function listRemoteTypingDaysFor(
-  uid: string,
-  machineHash: string,
-): Promise<UtcDay[]> {
-  const credentials = await requireSyncCredentials()
-  if (!credentials.ok) return []
-  const remoteFiles = await listFiles()
-  const perUid = collectRemoteOwnHashDays(remoteFiles, machineHash)
-  const days = perUid.get(uid)
-  if (!days) return []
-  return Array.from(days.keys()).sort()
-}
-
-/** Delete the cloud copy of a specific (uid, machineHash, day) and
- * its local mirror if we previously downloaded it. Used by the Sync >
- * Typing > Device > Delete-day UX: another device's record is gone
- * from cloud, and when that device next syncs the reconcile pass will
- * see its `uploaded` entry without a cloud file and drop its own
- * local copy (rule 3). Own-hash cache rows are accepted as stale until
- * the next rebuild — they live in the machine that owns the day.
- * Returns `true` when a cloud delete actually ran, `false` when the
- * user is unauthenticated or the cloud file was already missing. */
-export async function deleteRemoteTypingDay(
-  uid: string,
-  machineHash: string,
-  utcDay: UtcDay,
-): Promise<boolean> {
-  const credentials = await requireSyncCredentials()
-  if (!credentials.ok) return false
-  const remoteFiles = await listFiles()
-  const targetName = driveFileName(typingAnalyticsDeviceDaySyncUnit(uid, machineHash, utcDay))
-  const remoteFile = remoteFiles.find((f) => f.name === targetName)
-  const userData = app.getPath('userData')
-  try {
-    await unlink(deviceDayJsonlPath(userData, uid, machineHash, utcDay))
-  } catch (err) {
-    if ((err as NodeJS.ErrnoException).code !== 'ENOENT') {
-      log('warn', `typing-analytics local delete failed for ${uid} ${machineHash} ${utcDay}: ${String(err)}`)
-    }
-  }
-  // Tombstone the remote hash's cache rows for this day so the Data
-  // modal list refreshes immediately after the delete. Scoped to the
-  // single hash + day so a same-day local contribution stays visible.
-  try {
-    const { startMs, endMs } = utcDayBoundaryMs(utcDay)
-    const updatedAt = Date.now()
-    getTypingAnalyticsDB().tombstoneRowsForUidHashInRange(uid, machineHash, startMs, endMs, updatedAt)
-  } catch (err) {
-    log('warn', `typing-analytics cache tombstone failed for ${uid} ${machineHash} ${utcDay}: ${String(err)}`)
-  }
-  if (!remoteFile) return false
-  await deleteFile(remoteFile.id)
-  return true
-}
-
-/** Lazily fetch a single remote (uid, machineHash, day) into the
- * local cache. Returns `true` when the day was downloaded and merged,
- * `false` when the cloud copy was missing or a credential check failed.
- * Designed for the Sync > Typing > Device lazy-expand flow so the UI
- * can pull in only the days the user actually opens. */
-export async function fetchRemoteTypingDay(
-  uid: string,
-  machineHash: string,
-  utcDay: UtcDay,
-): Promise<boolean> {
-  const credentials = await requireSyncCredentials()
-  if (!credentials.ok) return false
-  const { password } = credentials
-  const remoteFiles = await listFiles()
-  const targetName = driveFileName(typingAnalyticsDeviceDaySyncUnit(uid, machineHash, utcDay))
-  const file = remoteFiles.find((f) => f.name === targetName)
-  if (!file) return false
-  const envelope = await downloadFile(file.id)
-  const plaintext = await decrypt(envelope, password)
-  const remoteBundle = JSON.parse(plaintext) as SyncBundle
-  const userData = app.getPath('userData')
-  const ownHash = await getMachineHash()
-  await mergeDeviceDayBundle(remoteBundle, { uid, machineHash, utcDay }, userData, ownHash)
-  return true
-}
-
-async function executeUploadSync(
-  password: string,
-  prefetchedFiles?: DriveFile[],
-  scope: SyncScope = 'all',
-): Promise<string[]> {
-  const remoteFilesInitial = prefetchedFiles ?? await listFiles()
-  // Run own-hash typing-analytics reconcile before collecting units so
-  // deleted cloud days don't get re-uploaded and vice-versa. The
-  // reconcile only deletes when it detects a divergence; when nothing
-  // changed we reuse the initial snapshot to keep the N+1 invariant.
-  let mutatedDuringReconcile = false
-  try {
-    const ownHash = await getMachineHash()
-    const result = await reconcileOwnHashTypingAnalytics(
-      remoteFilesInitial,
-      app.getPath('userData'),
-      ownHash,
-    )
-    mutatedDuringReconcile = result.mutated
-  } catch (err) {
-    log('warn', `typing-analytics reconcile failed: ${String(err)}`)
-  }
-  let syncUnits = await collectAllSyncUnits()
-  if (scope !== 'all') {
-    syncUnits = syncUnits.filter((unit) => matchesScope(unit, scope))
-  }
-  const remoteFiles = mutatedDuringReconcile ? await listFiles() : remoteFilesInitial
-  updateRemoteState(remoteFiles)
-  const total = syncUnits.length
-  let completed = 0
-  const failedUnits: string[] = []
-  const limit = pLimit(SYNC_CONCURRENCY)
-
-  await Promise.allSettled(
-    syncUnits.map((syncUnit) =>
-      limit(async () => {
-        completed++
-        emitProgress({
-          direction: 'upload',
-          status: 'syncing',
-          syncUnit,
-          current: completed,
-          total,
-        })
-
-        try {
-          await syncOrUpload(syncUnit, password, remoteFiles)
-        } catch (err) {
-          failedUnits.push(syncUnit)
-          emitProgress({
-            direction: 'upload',
-            status: 'error',
-            syncUnit,
-            message: errorMessage(err, 'Upload failed'),
-          })
-        }
-      }),
-    ),
-  )
-
-  // Refresh remote state once after all uploads to prevent polling re-downloads
-  const updatedFiles = await listFiles()
-  updateRemoteState(updatedFiles)
-
-  return failedUnits
-}
-
-// --- Remote state tracking ---
-
-function updateRemoteState(files: DriveFile[]): void {
-  lastKnownRemoteState.clear()
-  for (const file of files) {
-    lastKnownRemoteState.set(file.name, file.modifiedTime)
-  }
-}
-
-// --- Polling ---
-
-async function pollForRemoteChanges(): Promise<void> {
-  if (isSyncing) return
-  isSyncing = true
-
-  try {
-    const credentials = await requireSyncCredentials()
-    if (!credentials.ok) return  // polling stays silent — manual sync surfaces the reason
-    const password = credentials.password
-
-    const remoteFiles = await listFiles()
-
-    if (!passwordCheckValidated) {
-      await validatePasswordCheck(password, remoteFiles)
-    }
-
-    // First poll: just validate password and record remote state
-    // Avoids downloading all files on startup
-    if (lastKnownRemoteState.size === 0) {
-      updateRemoteState(remoteFiles)
-      return
-    }
-
-    const localKeyboardUids = await listLocalKeyboardUids()
-    // {file, syncUnit} pairs resolved once here rather than re-parsing
-    // the filename again inside the merge loop below.
-    const changedFiles = remoteFiles.flatMap((file) => {
-      if (lastKnownRemoteState.get(file.name) === file.modifiedTime) return []
-      const syncUnit = syncUnitFromFileName(file.name)
-      if (!syncUnit) return []
-      // analytics: handled by executeAnalyticsSync (Analyze panel mount).
-      if (isAnalyticsSyncUnit(syncUnit)) return []
-      // run logs: no dedicated on-demand sync entry point yet (see
-      // isRunLogSyncUnit's doc comment) — before-quit flush and manual
-      // sync still cover it, 3-minute polling does not.
-      if (isRunLogSyncUnit(syncUnit)) return []
-      if (!shouldDownloadSyncUnit(syncUnit, 'all', localKeyboardUids)) return []
-      return [{ file, syncUnit }]
-    })
-
-    updateRemoteState(remoteFiles)
-
-    const limit = pLimit(SYNC_CONCURRENCY)
-    const failedUnits: string[] = []
-    await Promise.allSettled(
-      changedFiles.map(({ file: remoteFile, syncUnit }) =>
-        limit(async () => {
-          try {
-            await mergeWithRemote(remoteFile, syncUnit, password, remoteFiles)
-            emitProgress({
-              direction: 'download',
-              status: 'success',
-              syncUnit,
-              message: 'Sync complete',
-            })
-          } catch (err) {
-            failedUnits.push(syncUnit)
-            if (err instanceof MalformedSyncBundleError) {
-              // Contained per-unit, same as any other poll failure — but
-              // deliberately do NOT forget the modifiedTime below: this
-              // poll already recorded it into lastKnownRemoteState above
-              // (before this loop ran), so leaving it in place makes the
-              // NEXT poll see this exact revision as unchanged and skip
-              // retrying it forever. A later fix (a new modifiedTime) is
-              // a different key and is retried normally. Manual syncs
-              // (executeDownloadSync/executeUploadSync) have no such
-              // memory and may retry a malformed unit on every attempt —
-              // deliberate, since those are visible, user-initiated
-              // actions, not a silent background loop. Unit name only —
-              // never bundle content — per the project's
-              // no-payload-in-logs rule for attacker-reachable remote data.
-              log('warn', `sync: ${err.message}`)
-              return
-            }
-            // Forget the modifiedTime so the next poll re-detects this file as changed
-            // and gets another chance to merge it.
-            lastKnownRemoteState.delete(remoteFile.name)
-          }
-        }),
-      ),
-    )
-
-    // Pass-level GC — see pack-gc.ts's doc for why this must never be
-    // triggered from inside a single unit's own merge callback above, and
-    // for why `failedUnits` is passed separately (skips that store's
-    // sweep only, not the whole GC call).
-    await runPackGcAfterPass(changedFiles.map((f) => f.syncUnit), failedUnits)
-  } catch {
-    // Polling failed — will retry next interval
-  } finally {
-    isSyncing = false
-  }
-}
-
-export function startPolling(): void {
-  if (pollTimer) return
-  pollTimer = setInterval(() => {
-    void pollForRemoteChanges()
-  }, POLL_INTERVAL_MS)
-}
-
-export function stopPolling(): void {
-  if (pollTimer) {
-    clearInterval(pollTimer)
-    pollTimer = null
-  }
-}
-
-// --- Analyze-panel analytics sync ---
-
-/** Per-uid mutex so switching between keyboards while the previous
- * sync is still running doesn't immediately skip the new uid. Uses
- * a Set instead of a single flag so `uid-a` and `uid-b` can proceed
- * in parallel — the cloud file namespace (`keyboards/{uid}/devices/*`)
- * is disjoint across uids so there is no conflict. */
-const analyticsSyncingUids = new Set<string>()
-
-/** Pull + push typing-analytics bundles for one keyboard, triggered
- * from the Analyze panel mount. Runs on its own per-uid mutex so
- * polling / manual sync stay untouched — the cloud file namespace is
- * disjoint (only `keyboards/{uid}/devices/*` is written) so there is
- * no conflict with the global `isSyncing` path.
- *
- * Returns true on a fully-successful pass so the caller can stamp a
- * rate-limit timestamp; returns false on skip (this uid is already
- * syncing or credentials are missing) or on any per-unit failure so
- * the caller can retry on the next Analyze mount. */
-export async function executeAnalyticsSync(uid: string): Promise<boolean> {
-  if (analyticsSyncingUids.has(uid)) return false
-  analyticsSyncingUids.add(uid)
-  try {
-    const credentials = await requireSyncCredentials()
-    if (!credentials.ok) return false
-    const password = credentials.password
-
-    // Drive-side prefix filter: scope the listing to this keyboard's
-    // analytics files. The in-memory `isAnalyticsSyncUnit` + `startsWith`
-    // checks below remain as a safety net in case Drive ever returns a
-    // looser substring match.
-    const prefix = `keyboards/${uid}/devices/`
-    const remoteFiles = await listFiles({ nameContains: driveFilenamePrefix(prefix) })
-    let anyFailure = false
-    const limit = pLimit(SYNC_CONCURRENCY)
-    // Units `mergeWithRemote` already handled — it uploads any
-    // divergence internally, so the push pass can skip them.
-    const mergedUnits = new Set<string>()
-
-    await Promise.allSettled(
-      remoteFiles.map((file) =>
-        limit(async () => {
-          const unit = syncUnitFromFileName(file.name)
-          if (!unit || !isAnalyticsSyncUnit(unit)) return
-          if (!unit.startsWith(prefix)) return
-          try {
-            await mergeWithRemote(file, unit, password, remoteFiles)
-            mergedUnits.add(unit)
-          } catch {
-            anyFailure = true
-          }
-        }),
-      ),
-    )
-
-    const localUnits = await collectAnalyticsSyncUnitsForUid(uid)
-    await Promise.allSettled(
-      localUnits
-        .filter((unit) => !mergedUnits.has(unit))
-        .map((unit) =>
-          limit(async () => {
-            try {
-              await syncOrUpload(unit, password, remoteFiles)
-            } catch {
-              anyFailure = true
-            }
-          }),
-        ),
-    )
-
-    return !anyFailure
-  } catch {
-    return false
-  } finally {
-    analyticsSyncingUids.delete(uid)
-  }
-}
-
-// --- Debounced upload ---
-
-export function notifyChange(syncUnit: string): void {
-  pendingChanges.add(syncUnit)
-  broadcastPendingStatus()
-
-  if (debounceTimer) {
-    clearTimeout(debounceTimer)
-  }
-
-  debounceTimer = setTimeout(() => {
-    void flushPendingChanges()
-  }, DEBOUNCE_MS)
-}
-
-async function flushPendingChanges(): Promise<void> {
-  if (pendingChanges.size === 0) return
-
-  if (isSyncing) {
-    debounceTimer = setTimeout(() => {
-      void flushPendingChanges()
-    }, DEBOUNCE_MS)
-    return
-  }
-
-  isSyncing = true
-
-  debounceTimer = null
-
-  try {
-    const config = await loadAppConfig()
-    if (!config.autoSync) {
-      pendingChanges.clear()
-      broadcastPendingStatus()
-      return
-    }
-
-    const credentials = await requireSyncCredentials()
-    if (!credentials.ok) {
-      pendingChanges.clear()
-      broadcastPendingStatus()
-      return
-    }
-    const password = credentials.password
-
-    const changes = new Set(pendingChanges)
-    pendingChanges.clear()
-
-    emitProgress({ direction: 'upload', status: 'syncing', message: 'Auto-sync starting...' })
-
-    const remoteFiles = await listFiles()
-    updateRemoteState(remoteFiles)
-
-    if (!passwordCheckValidated) {
-      try {
-        await validatePasswordCheck(password, remoteFiles)
-      } catch (err) {
-        for (const unit of changes) pendingChanges.add(unit)
-        broadcastPendingStatus()
-        if (err instanceof PasswordMismatchError) {
-          emitProgress({ direction: 'upload', status: 'error', message: 'sync.passwordMismatch' })
-        } else {
-          emitProgress({ direction: 'upload', status: 'error', message: errorMessage(err, 'Password check failed') })
-        }
-        return
-      }
-    }
-
-    const limit = pLimit(SYNC_CONCURRENCY)
-    await Promise.allSettled(
-      [...changes].map((syncUnit) =>
-        limit(async () => {
-          try {
-            await syncOrUpload(syncUnit, password, remoteFiles)
-          } catch {
-            // Re-add failed unit so pending stays true
-            pendingChanges.add(syncUnit)
-          }
-        }),
-      ),
-    )
-
-    broadcastPendingStatus()
-
-    // Refresh remote state after uploads to prevent polling re-downloads
-    const updatedFiles = await listFiles()
-    updateRemoteState(updatedFiles)
-
-    if (pendingChanges.size === 0) {
-      emitProgress({ direction: 'upload', status: 'success', message: 'Sync complete' })
-    } else {
-      emitProgress({ direction: 'upload', status: 'error', message: 'Some sync units failed' })
-    }
-  } finally {
-    isSyncing = false
-  }
-}
-
-// --- Before-quit handler ---
-
-interface BeforeQuitFinalizer {
-  hasWork: () => boolean
-  run: () => Promise<void>
-}
-
-const preSyncFinalizers: BeforeQuitFinalizer[] = []
-const extraFinalizers: BeforeQuitFinalizer[] = []
-
-/**
- * Register a finalizer that runs BEFORE the sync flush at before-quit time.
- * Use this when the subsystem's flush may enqueue new sync units via
- * notifyChange() — running pre-sync guarantees the freshly queued units land
- * in the same quit cycle instead of waiting for the next launch.
- */
-export function registerPreSyncQuitFinalizer(finalizer: BeforeQuitFinalizer): void {
-  preSyncFinalizers.push(finalizer)
-}
-
-/**
- * Register an additional async finalizer to run alongside the sync flush at
- * before-quit time. Used by subsystems that do not touch the sync queue.
- */
-export function registerBeforeQuitFinalizer(finalizer: BeforeQuitFinalizer): void {
-  extraFinalizers.push(finalizer)
-}
-
-export function setupBeforeQuitHandler(): void {
-  app.on('before-quit', (e) => {
-    if (isQuitting) return
-
-    stopPolling()
-
-    const syncPending = pendingChanges.size > 0 || debounceTimer !== null
-    const preSync = preSyncFinalizers.filter((f) => f.hasWork())
-    const extras = extraFinalizers.filter((f) => f.hasWork())
-    if (!syncPending && preSync.length === 0 && extras.length === 0) return
-
-    e.preventDefault()
-    isQuitting = true
-
-    if (debounceTimer) {
-      clearTimeout(debounceTimer)
-      debounceTimer = null
-    }
-
-    const runQuitPhases = async (): Promise<void> => {
-      // Phase 1: pre-sync finalizers. They may call notifyChange() to
-      // enqueue additional sync units; those land in pendingChanges before
-      // the sync flush starts.
-      if (preSync.length > 0) {
-        await Promise.all(
-          preSync.map((f) =>
-            f.run().catch((err: unknown) => {
-              log('error', `pre-sync quit finalizer failed: ${String(err)}`)
-            }),
-          ),
-        )
-      }
-
-      // Phase 2: sync flush. Re-evaluate pendingChanges because pre-sync
-      // finalizers may have added to it.
-      if (syncPending || pendingChanges.size > 0) {
-        await flushPendingChanges().catch((err: unknown) => {
-          log('error', `before-quit sync flush failed: ${String(err)}`)
-        })
-      }
-
-      // Phase 3: remaining extra finalizers. Re-check hasWork() so nothing
-      // is run twice if it also happens to sit on the extra list.
-      const extrasAfter = extraFinalizers.filter((f) => f.hasWork())
-      if (extrasAfter.length > 0) {
-        await Promise.all(
-          extrasAfter.map((f) =>
-            f.run().catch((err: unknown) => {
-              log('error', `extra quit finalizer failed: ${String(err)}`)
-            }),
-          ),
-        )
-      }
-    }
-
-    // Always call app.quit() even if a phase unexpectedly throws, so the
-    // app cannot hang on the preventDefault()'d quit.
-    runQuitPhases()
-      .catch((err: unknown) => {
-        log('error', `before-quit phases crashed: ${String(err)}`)
-      })
-      .finally(() => {
-        app.quit()
-      })
-  })
-}
-
-// --- Test helpers ---
-
-export function _resetForTests(): void {
-  if (debounceTimer) {
-    clearTimeout(debounceTimer)
-    debounceTimer = null
-  }
-  stopPolling()
-  pendingChanges.clear()
-  lastKnownRemoteState.clear()
-  isSyncing = false
-  isQuitting = false
-  progressCallback = null
-  passwordCheckValidated = false
-  preSyncFinalizers.length = 0
-  extraFinalizers.length = 0
-}
+export { listUndecryptableFiles, scanRemoteData, fetchRemoteBundle, listRemoteFileNames } from './sync-scan'
+
+export {
+  changePassword,
+  resetPasswordCheckCache,
+  checkPasswordCheckExists,
+  setPasswordAndValidate,
+} from './sync-password'
+
+export type { SyncExecuteResult } from './sync-execute'
+export { executeSync } from './sync-execute'
+
+export {
+  hasAnyRemoteTypingData,
+  listRemoteTypingHashesForUidFromCloud,
+  listRemoteTypingDaysFor,
+  deleteRemoteTypingDay,
+  fetchRemoteTypingDay,
+} from './sync-typing-remote'
+
+export { startPolling, stopPolling } from './sync-polling'
+
+export { executeAnalyticsSync } from './sync-analytics'
+
+export {
+  notifyChange,
+  registerPreSyncQuitFinalizer,
+  registerBeforeQuitFinalizer,
+  setupBeforeQuitHandler,
+} from './sync-flush'
